+++ src/lib.rs
@@ -3,3 +3,8 @@ mod properties;
 
 pub use element::*;
 pub use properties::*;
+
+#[cfg(feature = "group")]
+mod group;
+#[cfg(feature = "group")]
+pub use group::*;