+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-11-06
+
+### Added
+- Period property
+- Group property
+- Group symbol property
+- Group trivial name property
+- Example that prints a simple periodic table to the console
+
+### Changed
+- Element name and string return static strings
+
 ## [0.1.0] - 2022-11-05
 
 Initial release.